test: move iroh tests into separate module
